Add error handling with toasts
When user logs in, authentication is stored in localstorage and a interceptor
appends the Authentication header to each request going to the backend.

Effectively, the user seems logged-in but when a request is sent to the backend
and it fails to authenticate, the frontend re-displays the login page.

Updated the query.service.ts to handle the new /queries api

New file with routes, apiEndpoin.ts