Hide modifying buttons depending on the role of logged-in user
There are certain elements that the standard user should not be able
to access. Even though backend will block such unauthorized requests,
it is nice to not have a clutter of useless buttons for the end-user.

Also, The queries page was re-structured.
There is a problem adding a click event listener to a row and a event
listener for a button found inside said row. The click
propagates from the button to the row, effectivelly clicking in both
elements at the same time.
TODO: fix this